+++ runtime/common/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod constants;
+
+pub mod traits;
+pub use traits::{ identity };
