+++ runtime/identity/Cargo.toml
@@ -1,12 +1,16 @@
 [package]
-name = "polymesh-primitives"
-version = "0.6.0"
+name = "polymesh-runtime-identity"
+version = "0.1.0"
 authors = ["Polymath"]
 edition = "2018"
 
 [dependencies]
+polymesh-runtime-common = { package = "polymesh-runtime-common", path = "../common", default-features = false }
+polymesh-primitives = { package = "polymesh-primitives", path = "../../primitives", default-features = false }
+
 serde = { version = "1.0", optional = true, features = ["derive"] }
 codec = { package = "parity-scale-codec", version = "1.1.0", default-features = false, features = ["derive"] }
+
 primitives = { package = "substrate-primitives", git = "https://github.com/paritytech/substrate", default-features = false, rev = "7d7e74fb77b6bee2ce9d6ebafcae09caff2d0e50" }
 application-crypto = { package = "substrate-application-crypto", git = "https://github.com/paritytech/substrate", default-features = false, rev = "7d7e74fb77b6bee2ce9d6ebafcae09caff2d0e50" }
 substrate-client = { git = "https://github.com/paritytech/substrate", default-features = false, rev = "7d7e74fb77b6bee2ce9d6ebafcae09caff2d0e50" }
@@ -17,24 +21,20 @@ bitvec = { version = "0.14.0", default-features = false, features = ["alloc"] }
 babe = { package = "srml-babe", git = "https://github.com/paritytech/substrate", default-features = false, rev = "7d7e74fb77b6bee2ce9d6ebafcae09caff2d0e50" }
 srml-support = { git = "https://github.com/paritytech/substrate", default-features = false, rev = "7d7e74fb77b6bee2ce9d6ebafcae09caff2d0e50" }
 sr-io = { git = "https://github.com/paritytech/substrate", default-features = false, rev = "7d7e74fb77b6bee2ce9d6ebafcae09caff2d0e50" }
-hex = "0.4.0"
+timestamp = { package = "srml-timestamp", git = "https://github.com/paritytech/substrate", default-features = false, rev = "7d7e74fb77b6bee2ce9d6ebafcae09caff2d0e50" }
+system = { package = "srml-system", git = "https://github.com/paritytech/substrate", default-features = false, rev = "7d7e74fb77b6bee2ce9d6ebafcae09caff2d0e50" }
+
 ed25519-dalek = { git = "https://github.com/dalek-cryptography/ed25519-dalek.git", tag ="1.0.0-pre.2" }
 
 [patch.crates-io]
 ed25519-dalek = { git = "https://github.com/dalek-cryptography/ed25519-dalek.git", tag ="1.0.0-pre.2" }
-#curve25519-dalek = { git = "https://github.com/dalek-cryptography/curve25519-dalek.git", tag = "1.2.3" }
 
-#[replace]
-#"ed25519-dalek:1.0.0-pre.3" = { git = "https://github.com/dalek-cryptography/ed25519-dalek.git", tag ="1.0.0-pre.2" }
-
-
-[dev-dependencies]
-substrate-serializer = { git = "https://github.com/paritytech/substrate", rev = "7d7e74fb77b6bee2ce9d6ebafcae09caff2d0e50" }
-pretty_assertions = "0.5.1"
 
 [features]
 default = ["std"]
-std = [
+no_std = []
+only-staking = []
+std = [ 
 	"codec/std",
 	"primitives/std",
 	"substrate-client/std",
@@ -46,4 +46,5 @@ std = [
 	"srml-support/std",
 	"babe/std",
 	"sr-io/std",
+	"timestamp/std",
 ]
