+++ runtime/identity/src/lib.rs
@@ -42,18 +42,21 @@
 //! # TODO
 //!  - KYC is mocked: see [has_valid_kyc](./struct.Module.html#method.has_valid_kyc)
 
-use rstd::{convert::TryFrom, prelude::*};
-
-use crate::{balances, constants::did::USER};
-use primitives::{
+use polymesh_runtime_common::{
+    identity::Trait,
+    constants::did::USER
+};
+use polymesh_primitives::{
     Identity as DidRecord, IdentityId, Key, Permission, PreAuthorizedKeyInfo, Signer, SignerType,
     SigningItem,
 };
 
 use codec::Encode;
 use core::convert::From;
+
+use rstd::{convert::TryFrom, prelude::*};
 use sr_io::blake2_256;
-use sr_primitives::{
+use runtime_primitives::{
     traits::{Dispatchable, Verify},
     AnySignature, DispatchError,
 };
@@ -64,100 +67,11 @@ use srml_support::{
     traits::{Currency, ExistenceRequirement, WithdrawReason},
     Parameter,
 };
-use substrate_primitives::{
+use primitives::{
     sr25519::{Public, Signature},
     H512,
 };
 use system::{self, ensure_signed};
-
-#[derive(codec::Encode, codec::Decode, Default, Clone, PartialEq, Eq, Debug)]
-pub struct Claim<U> {
-    issuance_date: U,
-    expiry: U,
-    claim_value: ClaimValue,
-}
-
-#[derive(codec::Encode, codec::Decode, Default, Clone, PartialEq, Eq, Debug)]
-pub struct ClaimMetaData {
-    claim_key: Vec<u8>,
-    claim_issuer: IdentityId,
-}
-
-#[derive(codec::Encode, codec::Decode, Default, Clone, PartialEq, Eq, Debug)]
-pub struct ClaimValue {
-    pub data_type: DataTypes,
-    pub value: Vec<u8>,
-}
-
-#[derive(codec::Encode, codec::Decode, Clone, Copy, PartialEq, Eq, Debug, PartialOrd, Ord)]
-pub enum DataTypes {
-    U8,
-    U16,
-    U32,
-    U64,
-    U128,
-    Bool,
-    VecU8,
-}
-
-impl Default for DataTypes {
-    fn default() -> Self {
-        DataTypes::VecU8
-    }
-}
-
-/// Keys could be linked to several identities (`IdentityId`) as master key or signing key.
-/// Master key or external type signing key are restricted to be linked to just one identity.
-/// Other types of signing key could be associated with more that one identity.
-#[derive(codec::Encode, codec::Decode, Clone, PartialEq, Eq, Debug)]
-pub enum LinkedKeyInfo {
-    Unique(IdentityId),
-    Group(Vec<IdentityId>),
-}
-
-pub type AuthorizationNonce = u64;
-
-/// It represents an authorization that any account could sing to allow operations related with a
-/// target identity.
-///
-/// # Safety
-///
-/// Please note, that `nonce` has been added to avoid **replay attack** and it should be the current
-/// value of nonce of master key of `target_id`. See `System::account_nonce`.
-/// In this way, the authorization is delimited to an specific transaction (usually the next one)
-/// of master key of target identity.
-#[derive(codec::Encode, codec::Decode, Clone, PartialEq, Eq, Debug)]
-pub struct TargetIdAuthorization<Moment> {
-    /// Target identity which is authorized to make an operation.
-    pub target_id: IdentityId,
-    /// It HAS TO be `target_id` authorization nonce: See `Identity::offchain_authorization_nonce`
-    pub nonce: AuthorizationNonce,
-    pub expires_at: Moment,
-}
-
-/// It is a signing item with authorization of that singning key (off-chain operation) to be added
-/// to an identity.
-/// `auth_signature` is the signature, generated by signing item, of `TargetIdAuthorization`.
-///
-/// # TODO
-///  - Replace `H512` type by a template type which represents explicitly the relation with
-///  `TargetIdAuthorization`.
-#[derive(codec::Encode, codec::Decode, Clone, PartialEq, Eq, Debug)]
-pub struct SigningItemWithAuth {
-    /// Signing item to be added.
-    pub signing_item: SigningItem,
-    /// Off-chain authorization signature.
-    pub auth_signature: H512,
-}
-
-/// The module's configuration trait.
-pub trait Trait: system::Trait + balances::Trait + timestamp::Trait {
-    /// The overarching event type.
-    type Event: From<Event<Self>> + Into<<Self as system::Trait>::Event>;
-    /// An extrinsic call.
-    type Proposal: Parameter + Dispatchable<Origin = Self::Origin>;
-}
-
 decl_storage! {
     trait Store for Module<T: Trait> as identity {
 
@@ -1019,7 +933,7 @@ impl<T: Trait> Module<T> {
     }
 
     /// It set/reset the current identity.
-    pub(super) fn set_current_did(did_opt: Option<IdentityId>) {
+    pub fn set_current_did(did_opt: Option<IdentityId>) {
         if let Some(did) = did_opt {
             <CurrentDid>::put(did);
         } else {
@@ -1053,17 +967,6 @@ impl<T: Trait> Module<T> {
             <PreAuthorizedJoinDid>::remove(signer);
         }
     }
-}
-
-pub trait IdentityTrait<T> {
-    fn get_identity(key: &Key) -> Option<IdentityId>;
-    fn is_signer_authorized(did: IdentityId, signer: &Signer) -> bool;
-    fn is_signer_authorized_with_permissions(
-        did: IdentityId,
-        signer: &Signer,
-        permissions: Vec<Permission>,
-    ) -> bool;
-    fn is_master_key(did: IdentityId, key: &Key) -> bool;
 }
 
 impl<T: Trait> IdentityTrait<T::Balance> for Module<T> {
