+++ runtime/runtime/Cargo.toml
@@ -56,6 +56,8 @@ authority-discovery-primitives = { package = "substrate-authority-discovery-prim
 srml-staking-reward-curve = { git = "https://github.com/paritytech/substrate", rev = "7d7e74fb77b6bee2ce9d6ebafcae09caff2d0e50" }
 phragmen = { package = "substrate-phragmen", git = "https://github.com/paritytech/substrate", default-features = false, rev = "7d7e74fb77b6bee2ce9d6ebafcae09caff2d0e50" }
 
+identity = { package = "polymesh-runtime-identity", path = "identity", default-features = false }
+
 [dev-dependencies]
 test-client = { package = "substrate-test-runtime-client", git = "https://github.com/paritytech/substrate", rev = "7d7e74fb77b6bee2ce9d6ebafcae09caff2d0e50" }
